Amounts should support big numbers

+++ src/me/Zrips/TradeMeAddon/Exp.java
@@ -178,7 +178,7 @@ public class Exp implements TradeModeInterface {
      */
     @Override
     public void Change(TradeInfo trade, int slot, ClickType button) {
-	int amount = amounts.get(slot);
+	Double amount = amounts.get(slot);
 	int PlayerExp = (int) getPlayerExperience(trade.getP1());
 	double OfferedExp = trade.getOffer(at);
